--- src/packages/2StatusBar/src/2Utils/index.ts
@@ -1,2 +1,0 @@
-export * from './2getPercentage';
-export * from './2getStatus';
