+++ src/packages/2StatusBar/src/Utils/getPercentage.ts
