+++ src/packages/2StatusBar/src/Utils/index.ts
@@ -1,0 +1,2 @@
+export * from './getPercentage';
+export * from './getStatus';
